refactor(e2e): refine e2e tests about networkchaos
